+++ src/DotRecast.Detour/DtNodePool.cs
@@ -37,22 +37,20 @@ namespace DotRecast.Detour
             m_map.Clear();
         }
 
-        public List<DtNode> FindNodes(long id)
+        public int FindNodes(long id, out List<DtNode> nodes)
         {
-            var hasNode = m_map.TryGetValue(id, out var nodes);
-            ;
-            if (nodes == null)
+            var hasNode = m_map.TryGetValue(id, out nodes);
+            if (hasNode)
             {
-                nodes = new List<DtNode>();
+                return nodes.Count;
             }
 
-            return nodes;
+            return 0;
         }
 
         public DtNode FindNode(long id)
         {
             var hasNode = m_map.TryGetValue(id, out var nodes);
-            ;
             if (nodes != null && 0 != nodes.Count)
             {
                 return nodes[0];