+++ CHANGELOG.md
@@ -7,8 +7,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased] - yyyy-mm-dd
 
 ### Added
-
+- Added DtNodePool.GetNode, FindNode, FindNodes tests
+ 
 ### Fixed
+- Fixed SOH issue in dtNavMeshQuery.Raycast
+- Fixed SOH issue in DtProximityGrid.QueryItems
 
 ### Changed